fix(survey): harden submission flow and validation (#63)
- Move crypto.subtle inside try-block (crash on non-HTTPS)
- Safe JSON parsing with Content-Type check (502/504 handling)
- Reject unknown answer keys in validateAnswers (data injection)
- Reject submissions for surveys without questions
- Add Sentry to detractor alert + top-level catch
- Add back button for multi-step navigation
- Fix freetext-required bypass (only optional skippable)
- Add exhaustive switch checks (default: never)
- Export RouteResult + noRouting() helper
- Remove unused practiceName prop
- Add .int() to Zod number validator

+++ src/lib/validations.ts
@@ -67,7 +67,7 @@ export const surveyResponseSchema = z.object({
   surveyId: z.string().uuid(),
   answers: z.record(
     z.string().max(100),
-    z.union([z.number().min(0).max(10), z.string().max(2000), z.boolean()])
+    z.union([z.number().int().min(0).max(10), z.string().max(2000), z.boolean()])
   ).refine(
     (obj) => Object.keys(obj).length <= 50,
     "Too many answers"
